dt-bindings: input: sun4i-lradc-keys: Add R329 and D1 compatibles
The R329 and D1 SoCs each contain an LRADC with a programming interface
compatible to earlier LRADCs. However, the LRADC now has its own clock
gate and reset line, instead of being always active.

To support this, add clock/reset properties to the binding, and require
them for the variant in the new SoCs.

Signed-off-by: Samuel Holland <samuel@sholland.org>