dt-bindings: hwlock: sun6i: Add interrupts property
While it was not officially documented until recently (e.g. A50), the
hwspinlock block can trigger an interrupt when a lock is unlocked. This
capability is used by Allwinner's ARISC firmware, it has been verified
to work on A64, and the IRQ numbers are reserved as far back as A31.
So most likely this feature has always been available.

Even though the Linux hwspinlock framework cannot make use of the IRQ,
the capability should still be documented in the device tree.

Signed-off-by: Samuel Holland <samuel@sholland.org>